#133 (CORE-3): detect Cyrillic (+ combining marks); dedupe the two GetScript copies
Script auto-detection had no branch for the Cyrillic block (U+0400-04FF), so Cyrillic
input fell through to Latin and was converted by Latn2Ipe — even though the
Convert(str, Script.Cyrillic) -> Cyrl2Deva branch already existed (dead, since detection
never returned Cyrillic). A Cyrillic Pāli query produced a wrong IPE term -> zero results.

- Add U+0400-04FF -> Cyrillic and U+0300-036F (combining diacritical marks used by the
  Cyrillic scheme) -> Unknown so they join the surrounding run instead of splitting it.
- Extract the duplicated GetScript from Any2Ipe and Any2Deva into a shared ScriptDetector
  so the two detection tables can't drift (they already had to be fixed in tandem for
  ZWJ/ZWNJ in #115). Any2Deva.GetScript stays a public delegator for existing callers/tests.

SRCH-9 note: Telugu/Thai/Tibetan/Khmer were assessed against the old CST4 Any2Ipe; in
CST.Core they are fully wired and work. Cyrillic was the only genuine gap.

Tests: Cyrillic classified and combining marks Unknown; a Cyrillic syllable converts to
the same IPE/Deva as its Devanagari equivalent; Thai regression confirms that parser is
wired. Conversion suite incl. #86 corpus oracle: 107/107.
